Move local-databases.ts to /databases/local-databases.ts
