+++ extensions/ql-vscode/src/data-extensions-editor/external-api-usage-query.ts
@@ -9,7 +9,7 @@ import {
 import { TeeLogger } from "../common";
 import { CancellationToken } from "vscode";
 import { CodeQLCliServer } from "../cli";
-import { DatabaseItem } from "../local-databases";
+import { DatabaseItem } from "../databases/local-databases";
 import { ProgressCallback } from "../progress";
 import { fetchExternalApiQueries } from "./queries";
 import { QueryResultType } from "../pure/new-messages";
