+++ extensions/ql-vscode/src/data-extensions-editor/generate-flow-model.ts
@@ -1,5 +1,5 @@
 import { CancellationToken } from "vscode";
-import { DatabaseItem } from "../local-databases";
+import { DatabaseItem } from "../databases/local-databases";
 import { join } from "path";
 import { QueryRunner } from "../query-server";
 import { CodeQLCliServer } from "../cli";
