+++ extensions/ql-vscode/src/databases/database-fetcher.ts
@@ -16,7 +16,7 @@ import { basename, join } from "path";
 import * as Octokit from "@octokit/rest";
 import { retry } from "@octokit/plugin-retry";
 
-import { DatabaseManager, DatabaseItem } from "../local-databases";
+import { DatabaseManager, DatabaseItem } from "./local-databases";
 import { showAndLogInformationMessage, tmpDir } from "../helpers";
 import { reportStreamProgress, ProgressCallback } from "../progress";
 import { extLogger } from "../common";
