+++ extensions/ql-vscode/src/extension.ts
@@ -43,7 +43,7 @@ import {
   TemplateQueryDefinitionProvider,
   TemplateQueryReferenceProvider,
 } from "./language-support";
-import { DatabaseManager } from "./local-databases";
+import { DatabaseManager } from "./databases/local-databases";
 import { DatabaseUI } from "./databases/local-databases-ui";
 import {
   DEFAULT_DISTRIBUTION_VERSION_RANGE,
