+++ extensions/ql-vscode/src/language-support/ast-viewer/ast-builder.ts
@@ -4,7 +4,7 @@ import {
   BqrsId,
   EntityValue,
 } from "../../pure/bqrs-cli-types";
-import { DatabaseItem } from "../../local-databases";
+import { DatabaseItem } from "../../databases/local-databases";
 import { ChildAstItem, AstItem } from "./ast-viewer";
 import { Uri } from "vscode";
 import { QueryOutputDir } from "../../run-queries-shared";
