+++ extensions/ql-vscode/src/language-support/ast-viewer/ast-viewer.ts
@@ -15,7 +15,7 @@ import {
 } from "vscode";
 import { basename } from "path";
 
-import { DatabaseItem } from "../../local-databases";
+import { DatabaseItem } from "../../databases/local-databases";
 import { UrlValue, BqrsId } from "../../pure/bqrs-cli-types";
 import { showLocation } from "../../interface-utils";
 import {
