+++ extensions/ql-vscode/src/local-queries/local-queries.ts
@@ -25,7 +25,7 @@ import { CoreCompletedQuery, QueryRunner } from "../query-server";
 import { QueryHistoryManager } from "../query-history/query-history-manager";
 import { DatabaseUI } from "../databases/local-databases-ui";
 import { ResultsView } from "../interface";
-import { DatabaseItem, DatabaseManager } from "../local-databases";
+import { DatabaseItem, DatabaseManager } from "../databases/local-databases";
 import {
   createInitialQueryInfo,
   getQuickEvalContext,
