+++ extensions/ql-vscode/src/query-server/legacy/upgrades.ts
@@ -10,7 +10,7 @@ import * as messages from "../../pure/legacy-messages";
 import * as qsClient from "./query-server-client";
 import * as tmp from "tmp-promise";
 import { dirname } from "path";
-import { DatabaseItem } from "../../local-databases";
+import { DatabaseItem } from "../../databases/local-databases";
 import { asError, getErrorMessage } from "../../pure/helpers-pure";
 import { redactableError } from "../../pure/errors";
