+++ extensions/ql-vscode/src/query-server/query-runner.ts
@@ -1,7 +1,7 @@
 import { CancellationToken } from "vscode";
 import { CodeQLCliServer } from "../cli";
 import { ProgressCallback } from "../progress";
-import { DatabaseItem } from "../local-databases";
+import { DatabaseItem } from "../databases/local-databases";
 import { QueryOutputDir } from "../run-queries-shared";
 import { Position, QueryResultType } from "../pure/new-messages";
 import { BaseLogger, Logger } from "../common";
