+++ extensions/ql-vscode/test/vscode-tests/minimal-workspace/local-databases.test.ts
@@ -11,7 +11,7 @@ import {
   DatabaseResolver,
   findSourceArchive,
   FullDatabaseOptions,
-} from "../../../src/local-databases";
+} from "../../../src/databases/local-databases";
 import { Logger } from "../../../src/common";
 import { ProgressCallback } from "../../../src/progress";
 import { CodeQLCliServer, DbInfo } from "../../../src/cli";
