+++ extensions/ql-vscode/test/vscode-tests/no-workspace/language-support/contextual/file-range-from-uri.test.ts
@@ -1,6 +1,6 @@
 import { Uri, Range } from "vscode";
 
-import { DatabaseItem } from "../../../../../src/local-databases";
+import { DatabaseItem } from "../../../../../src/databases/local-databases";
 import {
   WholeFileLocation,
   LineColumnLocation,
