+++ extensions/ql-vscode/test/vscode-tests/no-workspace/test-adapter.test.ts
@@ -12,7 +12,7 @@ import {
 
 import { QLTestAdapter } from "../../../src/test-adapter";
 import { CodeQLCliServer } from "../../../src/cli";
-import { DatabaseManager } from "../../../src/local-databases";
+import { DatabaseManager } from "../../../src/databases/local-databases";
 import { mockedObject } from "../utils/mocking.helpers";
 import { TestRunner } from "../../../src/test-runner";
 import {
